+++ src/main/java/service/interfaces/frontend/FeedbackService.java
@@ -20,5 +20,5 @@ public interface FeedbackService {
 
     List<FeedbackItem> findByCategory(String category);
 
-    double getAverageRating(String category);
+    int getAverageRating(String category);
 }
